requirements.txt
# Web Scraping
selenium==4.19.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3

# JSON, OS, etc. (modules intégrés à Python, pas besoin d'installation)
# json
# os
# time
# random
# re
# urllib.parse

# Base de données PostgreSQL
psycopg2-binary==2.9.9

# Traitement du langage naturel
langdetect==1.0.9
transformers==4.41.1
spacy==3.7.4
nltk==3.8.1
gensim==4.3.2
pandas==2.2.2

# Logging (intégré à Python)
# logging

# Téléchargements nécessaires (à exécuter séparément dans un script ou manuellement)
# python -m spacy download en_core_web_sm
# python -m spacy download fr_core_news_sm
# python -m spacy download es_core_news_sm
# python -m spacy download de_core_news_sm
# nltk.download('stopwords')
